release: 0.7.4
Search was our longest tooth; it had a lot of repeating code that
differed per implementation; we moved it into FDS in 0.7.1 but it still
had lots of unnecessary, repeating cruft and holdover props. It also was
a class component, not a functional one; we rewrite Search with hooks
and centralise all logic instead of repeating it to halve the length of
the component.